Add author, modified_time and expiration_time to og tags
This commit adds author type (as a profile with first_name, last_name, username
and gender fields) and modified_time and expiration_time as properties that the
user can set in their blog post's front matter. If set, the plugin will generate
og: tags for each of these.

I had to add the timezone to env.rb so the the tests always come out the same,
regardless of where you run them.